+++ src/components/navbar/index.ts
@@ -1,0 +1,2 @@
+export * from './AuthenticatedNavbar.tsx';
+export * from './BottomNavigationMobile.tsx';