+++ src/stores/userStore.ts
@@ -1,11 +1,11 @@
 import { UserWithAuth } from '@models/user.model.ts';
-import { UserStateZustand } from '@models/zustand.models.ts';
+import { UserStateZustand } from '@models/zustand.model';
 import { create, StateCreator } from 'zustand';
 import { devtools, persist, createJSONStorage } from 'zustand/middleware';
 
 const initialState = {
   username: null,
-  emial: null,
+  email: null,
   role: null,
   isAuthenticated: false,
 };
@@ -26,16 +26,8 @@ const userApi: StateCreator<UserStateZustand> = (set, get) => ({
 export const useUserStore = create<UserStateZustand>()(
   devtools(
     persist(userApi, {
-      name: 'user-storage', // name of the item in the storage (must be unique)
-      storage: createJSONStorage(() => sessionStorage), // (optional) by default, 'localStorage' is used
+      name: 'user-storage',
+      storage: createJSONStorage(() => sessionStorage),
     })
   )
 );
-
-// declare global {
-//   interface Window {
-//     store: typeof useUserStore;
-//   }
-// }
-//
-// window.store = useUserStore;