+++ src/utils/constants/index.ts
@@ -1,1 +1,4 @@
 export * from './api.constants.ts';
+export * from './categoriesTags.constants.ts';
+export * from './post.constants.ts';
+export * from './user.constants.ts';