+++ src/utils/constants/user.constants.ts
@@ -1,0 +1,15 @@
+import { LoginFormErrorValues } from '@models/error.model';
+import { UserSignUp } from '@models/user.model';
+
+export const initialValuesSignUp: UserSignUp = {
+  username: '',
+  email: '',
+  password: '',
+  confirmPasswd: '',
+  role: '',
+};
+
+export const initialValuesLogin: LoginFormErrorValues = {
+  nickname: '',
+  passwd: '',
+};