+++ src/utils/errors/index.ts
@@ -1,1 +1,1 @@
-export * from './group-errors.ts';
+export * from './group-errors.error.ts';